Fix compiler warnings: implicit conversion loses integer precision
src/utils_str.c:1250:15: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
        nbre_bytes = strlen (new_text);
                   ~ ^~~~~~~~~~~~~~~~~
src/utils_str.c:1251:15: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        nbre_chars = g_utf8_strlen (new_text, -1);
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/utils_str.c
@@ -1240,8 +1240,8 @@ gchar *utils_str_remove_accents (const gchar *text)
     gchar *tmp_str;
     gunichar ch;
     gint i = 0;
-    gint nbre_bytes;
-	gint nbre_chars;
+    size_t nbre_bytes;
+    size_t nbre_chars;
 
 	if (!text)
 		return NULL;
